GP-71 Optimize the code structure
